Ubuntu build HOWTO: Suggest to install git instead of git-core
The package docs for the git-core Ubuntu package in all supported
Ubuntu versions, including 12.04 say:

This is a transitional dummy package.  The 'git-core' package has been
renamed to 'git', which has been installed automatically.  This git-core
package is now obsolete, and can safely be removed from the system if no
other package depends on it.

Fixes #4109